% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{change_swap_par}
\alias{change_swap_par}
\title{renamed to: "change_swap_parameter()"}
\usage{
change_swap_par(param, name, value, verbose = F)
}
\arguments{
\item{param}{renamed to: "change_swap_parameter()"}

\item{name}{renamed to: "change_swap_parameter()"}

\item{value}{renamed to: "change_swap_parameter()"}

\item{verbose}{renamed to: "change_swap_parameter()"}
}
\value{
renamed to: "change_swap_parameter()"
}
\description{
renamed to: "change_swap_parameter()"
}
