% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{change_swap_table}
\alias{change_swap_table}
\title{Change SWAP Table}
\usage{
change_swap_table(table, variable, row, value, verbose = F)
}
\arguments{
\item{table}{can be one SWAP table, or a list of SWAP tables as loaded by
\code{load_swap_tables()}.}

\item{variable}{the name of the variable you would like to alter (string)}

\item{row}{the row you would like to alter (integer)}

\item{value}{the value you would like to enter (must be SWAP-FORTRAN compatible)}

\item{verbose}{print status? (flag)}
}
\value{
Returns the same table / list of tables passed to the function, with
the desired variable modified at the given row.
}
\description{
changes the value of a variable in the given row of a SWAP table.
}
