% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{change_swap_vector}
\alias{change_swap_vector}
\title{Change SWAP vector}
\usage{
change_swap_vector(vector, index, value, variable = NULL, verbose = F)
}
\arguments{
\item{vector}{dataframe of vector to alter. (Optionally, one can pass a list
of vectors as returned by \code{load_swap_vectors()} and then also pass the variable name)}

\item{index}{index of the vector to alter (integer)}

\item{value}{value to set the vector at the given index (SWAP-FORTRAN compatible)}

\item{variable}{optional, only required if passing a list of multiple vectors. (string)}
}
\value{
the same vector or vector list as passed, but with the modified value
}
\description{
Alters a SWAP vector with the given value at the given index. Can optionally
pass a list of vectors as returning by \code{load_swap_vectors()}, and the variable
name, and rswap will find the right vector for you
}
