% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{clean_swap_file}
\alias{clean_swap_file}
\title{Clean SWAP File}
\usage{
clean_swap_file(project_path, swap_file = "swap.swp")
}
\arguments{
\item{project_path}{path to the project directory (string)}

\item{swap_file}{name of the swap file (string)}
}
\value{
Returns a cleaned up SWAP file in string vector form.
}
\description{
Returns a cleaned up SWAP file fit for manipulation (removes all non-vital
information).
}
\details{
This function is designed to be used internally, but may be useful to the end
users. This is currently only tested for the SWAP main file.
}
