% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{clean_swp_file}
\alias{clean_swp_file}
\title{renamed to: "clean_swap_file"}
\usage{
clean_swp_file(project_path, swap_file = "swap.swp")
}
\arguments{
\item{project_path}{renamed to: "clean_swap_file"}

\item{swap_file}{renamed to: "clean_swap_file"}
}
\value{
renamed to: "clean_swap_file"
}
\description{
renamed to: "clean_swap_file"
}
