% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{convert_swap_humidity}
\alias{convert_swap_humidity}
\title{Convert RH to AVP}
\usage{
convert_swap_humidity()
}
\value{
tbc
}
\description{
Converts the commonly used relative humidity (RH) to the SWAP required
actual vapour pressure.
}
\examples{
#tbc
}
