% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{est_avp}
\alias{est_avp}
\title{Estimate Actual Vapour Pressure}
\usage{
est_avp(tmin, tmax, rh)
}
\arguments{
\item{tmin}{numeric vector, in Celsius (C)}

\item{tmax}{numeric vector, in Celsius (C)}

\item{rh}{numeric vector, in \% (ie. 53.2)}
}
\value{
vector of AVP, rounded to 2 decimal placews
}
\description{
Converts relative humidity (RH) to actual vapour pressure (AVP) using an
estimation with maximum and minimum temperature. This function exists because
SWAP uses AVP whereas many datasets contain RH. Read more on \href{https://www.fao.org/4/x0490e/x0490e07.htm}{fao.org}
}
\details{
Note: if you would like to use this for hourly AVP, then just use the same
value for tmin and tmax
}
\examples{

est_avp(tmin = c(1.8, 1.9, 2.0), tmax =	c(6, 7, 8),
                   rh = c(60.2, 40.5, 90.2))

}
\seealso{
\code{\link[=load_swap_meteo]{load_swap_meteo()}} \code{\link[=write_swap_meteo]{write_swap_meteo()}}
}
