% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_observed}
\alias{filter_observed}
\title{filter observed data}
\usage{
filter_observed(observed_data, var = NULL, depth = NULL, addtional = NULL)
}
\arguments{
\item{observed_data}{observed data as given by load_observed()}

\item{var}{\strong{OPT} name(s) of the variables you would like to select (string). leave blank for "all"}

\item{depth}{\strong{OPT} value(s) of the depths you would like to select (numeric). leave blank for "all"}

\item{addtional}{\strong{OPT} if you would like to select a specific column(s), enter
them here as a string/vector.}
}
\value{
dataframe consisting of DATE column, and desired observed values
}
\description{
by variable and depth
}
