% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{filter_swap_data}
\alias{filter_swap_data}
\title{Filter SWAP Data}
\usage{
filter_swap_data(data, var = NULL, depth = NULL)
}
\arguments{
\item{data}{as given by \code{load_observed()$data} or \code{read_swap_output()$custom_depth} (dataframe)}

\item{var}{name(s) of the variables you would like to select, leave blank for all. (string)}

\item{depth}{value(s) of the depths you would like to select, leave blank for all. (numeric)}
}
\value{
Returns a dataframe with a \code{DATE} column, followed by the desired data.
}
\description{
This is an internal function which filters dataframes of SWAP data by depth
and variable. It might be useful for end users as well, which is why its
accessible.
}
\examples{
# path to sample results
example_path <- "./inst/extdata/rswap_example_output"

# load some SWAP data (either observed, or modeled using read_swap_output())
data <- load_swap_observed(project_path = example_path, verbose = T)

filter_swap_data(data$data, var = "WC", depth = "15")

}
