% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_swap_table}
\alias{format_swap_table}
\title{Format SWAP Table}
\usage{
format_swap_table(table, verbose = FALSE)
}
\arguments{
\item{table}{dataframe of your SWAP table}

\item{verbose}{(flag) print status?}
}
\value{
A dataframe with formatted values ready to be written by \code{write_swap_tables()}
}
\description{
This handy function converts your dataframe into a SWAP compatible format
(mainly, instead of R writing 4 to disk, it writes 4.0 when its a double. Or
when its a string, it will write 'hupsel' instead of hupsel. )
}
\examples{

  ISUBLAY = c(1:5)
  ISOILLAY =  c(1:5)
  HSUBLAY =  c(15, 20, 5, 50, 50)
  HCOMP =  c(1, 5, 1, 5, 10)
  NCOMP =  HSUBLAY / HCOMP

  vdsp <- data.frame(ISUBLAY, ISOILLAY, HSUBLAY, HCOMP, NCOMP)


format_swap_table(vdsp)

}
\seealso{
\code{\link[=write_swap_tables]{write_swap_tables()}}
}
