% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_swap_vector}
\alias{format_swap_vector}
\title{Format SWAP Vector}
\usage{
format_swap_vector(vector, name, verbose = FALSE)
}
\arguments{
\item{vector}{a vector to be formatted for SWAP}

\item{name}{(string) name of vector to be formatted}

\item{verbose}{(flag) print status?}
}
\value{
A vector with formatted values ready to be written by \code{write_swap_vectors()}
}
\description{
This handy function converts your vector into a SWAP compatible format
(mainly, instead of R writing 4 to disk, it writes 4.0 when its a double. Or
when its a string, it will write 'hupsel' instead of hupsel.)
}
\examples{

format_swap_vector(c(10, 10, 10, 20, 30, 50), "DZNEW")
format_swap_vector(c("2003-12-31", "2004-12-31"), "OUTDAT")
format_swap_vector(c("2003-12-31", "2004-12-31"), "OUTDATINT")

}
\seealso{
\code{\link[=write_swap_vectors]{write_swap_vectors()}}
}
