% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_depths}
\alias{get_depths}
\title{get_depth}
\usage{
get_depths(data, variable = NULL)
}
\arguments{
\item{data}{\strong{REQ} (list) as given by load_observed()}

\item{variable}{\strong{OPT} (string) variable for which depth levels should be given.
If no variable is given, all depths will be returned}
}
\value{
numeric vector of depths
}
\description{
extract numeric depth values for given observed variable
}
