% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_swap_unit}
\alias{get_swap_unit}
\title{Get SWAP unit}
\usage{
get_swap_unit(variable)
}
\arguments{
\item{variable}{}
}
\value{
Returns unit of passed SWAP variable in string form.
}
\description{
Gets the unit of the a SWAP variable.
}
