% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_swap_units}
\alias{get_swap_units}
\title{get swap units}
\usage{
get_swap_units(variable)
}
\arguments{
\item{variable}{(REQ) (string) ("WC", "H","TEMP","DRAINAGE","RAIN")}
}
\value{
unit in string from
}
\description{
returns default swap units for the given variable
}
