% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_swap_variables}
\alias{get_swap_variables}
\title{Get SWAP variables}
\usage{
get_swap_variables(swap_data, verbose = F)
}
\arguments{
\item{swap_data}{dataframe as returned by other rswap functions.}

\item{verbose}{print status? (flag)}
}
\value{
Returns a character vector of detected variables
}
\description{
Gets the SWAP variable names from a SWAP dataframe as used by a variety of
rswap functions.
}
\seealso{
\code{get_swap_depths()}
\code{load_swap_observed()}
\code{load_swap_output()}
}
