% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_SWAPtools}
\alias{install_SWAPtools}
\title{Install SWAPtools}
\usage{
install_SWAPtools()
}
\value{
returns \code{TRUE} if package is already installed, or if installation
was successful. returns \code{FALSE} if installation failed.
}
\description{
Installs SWAPtools from \code{waterwijzerlandbouw.wur.nl/repo}
}
\keyword{internal}
