% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_last_run}
\alias{load_last_run}
\title{Load last run. Loads the most recently saved run of the project directory.}
\usage{
load_last_run(project_path, verbose = F)
}
\arguments{
\item{project_path}{String, path to the project directory / or custom save location.}

\item{verbose}{Logical. Prints status reports}

\item{custom_directory}{Logical, Set to TRUE if using custom save location.}
}
\description{
Load last run. Loads the most recently saved run of the project directory.
}
