% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_observed}
\alias{load_observed}
\title{Load observed data}
\usage{
load_observed(project_path, verbose = F)
}
\arguments{
\item{project_path}{String. Path to project directory.}

\item{verbose}{Logical. Prints status reports}
}
\description{
(make sure to use the template .xlsx file, placed in your project directory)
}
