% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{load_swap_meteo}
\alias{load_swap_meteo}
\title{Load SWAP meteo data}
\usage{
load_swap_meteo(project_path, verbose = FALSE)
}
\arguments{
\item{project_path}{Path to project directory (string)}

\item{verbose}{print status? (flag)}
}
\value{
A dataframe of the meteo data
}
\description{
This function loads the meteo file \code{rswap_meteo.csv} into the R-environment,
allowing to modify the contents at will. Descriptions of the variables and
their units are provided below.
}
\details{
\strong{NOTE:} This is my best "guess" at what the units of the meteo input file
are. I could not find any official definitions in the manual, so use with
caution and let me know if I got anything wrong.
\itemize{
\item \code{Station}  --  a string, name of station, ie. "'Robert'"
\item \code{DD}       -- day of month, ie. "01"
\item \code{MM}       -- month number, ie. "01"
\item \code{YYYY}     -- year, ie "2001"
\item \code{Rad}      -- solar radiation, unit is  kJ/m^2 ie "3810.1"
\item \code{Tmin}     -- Minimum temperature (daily) in degrees Celsius, ie "-3.2"
\item \code{Tmax}     -- Maximum temperature (daily) in degrees Celsius, ie. "5.3"
\item \code{Hum}      -- Humidity (kPa?) (See note below) ie. "0.524"
\item \code{Wind}     -- Wind speed (m/s) ie. "4.5"
\item \code{Rain}     -- precipitation, (mm) ie. "0.2"
\item \code{ETref}    -- (optional) Reference Evapotranspiration (mm) ie. "0.4"
\item \code{Wet}      -- (optional) Rain intensity, fraction of day between 0 and 1, ie. "0.029"
}
}
\examples{
#tbc
}
