% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{load_swap_parameters}
\alias{load_swap_parameters}
\title{Load SWAP parameters}
\usage{
load_swap_parameters(project_path, file_type = ".swp", verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string}

\item{file_type}{type of swap file to load parameters from (either '.swp' for main file or '.dra' for drain file)}

\item{verbose}{print status? (flag)}
}
\value{
Tibble of parameter set
}
\description{
Loads the SWAP parameters as parsed by \code{parse_swap_file()} and returns them as
a tibble. Requires a parsed file first!
}
\seealso{
\code{\link[=parse_swap_file]{parse_swap_file()}} \code{\link[=write_swap_file]{write_swap_file()}}
}
