% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{load_swap_tables}
\alias{load_swap_tables}
\title{Load SWAP tables}
\usage{
load_swap_tables(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{Path to project directory (string)}

\item{swap_file}{name of the swap file to parse (string)}

\item{verbose}{print status? (flag)}
}
\value{
Returns list of tibbles named by their column names.
}
\description{
This function loads all the SWAP tables as parsed by \code{parse_swap_file()}If
the SWAP vectors have not been parsed, then this will be done first using
\code{swap_file}
}
\seealso{
\code{\link[=parse_swap_file]{parse_swap_file()}} \code{\link[=write_swap_file]{write_swap_file()}}
}
