% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{load_swap_vectors}
\alias{load_swap_vectors}
\title{Load SWAP Vectors}
\usage{
load_swap_vectors(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{path to project_directory (string)}

\item{swap_file}{name of the swap file to parse (string)}

\item{verbose}{print status? (flag)}
}
\value{
Returns a list of vectors
}
\description{
This function loads all the SWAP vectors as parsed by \code{parse_swp_file()}
}
\details{
Something to consider: Currently the function returns the vectors as a normal
dataframe with the first row being "VARNAME = ". this is because this is the
format in which the vectors are written to be understood by SWAP. One could
return the vectors in a real vector format, but then one would have to do
some extra steps when writing them again. I'll consider what will be best, but
for now I'll stick to the dataframe version.
}
