% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_variables_db}
\alias{load_variables_db}
\title{Loads the SWAPtools variables database}
\usage{
load_variables_db()
}
\value{
returns \code{TRUE} if database has been loaded (for now just SWAPtools_variables.rds)
and \code{FALSE} if it has not been loaded.
}
\description{
which gives package wide access to the SWAPtools databases, and means
individual function calls do not need to re-load the database for their
individual environments.
}
