% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{match_mod_obs}
\alias{match_mod_obs}
\title{match mod obs}
\usage{
match_mod_obs(
  project_path,
  variable,
  depth = NULL,
  verbose = F,
  archived = FALSE
)
}
\arguments{
\item{project_path}{req}

\item{variable}{req}

\item{depth}{opt}

\item{verbose}{opt}

\item{archived}{}
}
\value{
list of modelled DF and observed DF, matched
}
\description{
internal function: matches observed and modelled dataframes and returns them
}
\keyword{internal}
