% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{match_swap_data}
\alias{match_swap_data}
\title{Match Modeled Values To Observed Values}
\usage{
match_swap_data(
  project_path,
  variable,
  depth = NULL,
  verbose = F,
  archived = F
)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to match (string)}

\item{depth}{an optional depth to select for (numeric)}

\item{verbose}{print status? (boolean)}

\item{archived}{is the project in /rswap_saved/?}
}
\value{
Returns a list of two dataframes, \code{.$mod} and \code{.$obs}. These two dataframes have identical dimensions and column names.
}
\description{
This function is used internally to match observed values with modeled values
in the same format, such that it is easy to pass them to performance indicators.
}
\details{
This function is available since it might be of use to the end user.
}
\examples{
# path to sample results
example_path <- "./inst/extdata/rswap_example_output"

match_swap_data(example_path, "WC", depth = 15, verbose = T, archived = F)
}
