% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{melt_all_runs}
\alias{melt_all_runs}
\title{renamed to: "melt_swap_runs"}
\usage{
melt_all_runs(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{renamed to: "melt_swap_runs"}

\item{variable}{renamed to: "melt_swap_runs"}

\item{depth}{renamed to: "melt_swap_runs"}

\item{verbose}{renamed to: "melt_swap_runs"}
}
\value{
renamed to: "melt_swap_runs"
}
\description{
renamed to: "melt_swap_runs"
}
