% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{melt_swap_data}
\alias{melt_swap_data}
\title{Melt SWAP data}
\usage{
melt_swap_data(project_path, variable = NULL, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory \code{string}}

\item{variable}{variables to include in the melt \code{string}}

\item{depth}{depths to include in the melt \code{numeric}}

\item{verbose}{print status? \code{flag}}
}
\value{
Melted dataframe of matched modeled and observed values
}
\description{
This function transforms your observed and modeled data to by in tidy format.
This makes it easier to use with other r packages such as \code{ggplot2}.
}
\details{
This function will adjust to any variables and depths you give it. If you do
not provide any variables, or any depths, all available ones will be used.
}
\seealso{
\code{\link[=match_swap_data]{match_swap_data()}}
}
