% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{melt_swap_runs}
\alias{melt_swap_runs}
\title{Melt All Runs}
\usage{
melt_swap_runs(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to be used (string)}

\item{depth}{optional depth to filter by, leave blank for all/no depth (numeric)}

\item{verbose}{print status? (flag)}
}
\value{
Returns dataframe with columns "run" "DATE" "tag" "variable" "value"
}
\description{
This function combines the modeled and observed data of the current run,
along with all the modeled and observed data of all the saved runs
(for a single variable) and returns them in tidy format, easy to use for
plotting, etc.
}
\details{
This is once again an internal function that is made available due to it
possibly being useful for the end user.
}
\examples{
# path to sample results (only for package internal example, you don't need
# this function)
example_path <- system.file(package = "rswap", "extdata/rswap_example_output")

# in your own projects, you would use something like this:
# example_path <- "C:/path/to/swap_folder/example_project"

melt_swap_runs(example_path, "WC", depth = 15, verbose = TRUE)
}
