% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{modify_swap_file}
\alias{modify_swap_file}
\title{Modify SWAP file}
\usage{
modify_swap_file(
  project_path,
  input_file = NULL,
  output_file = NULL,
  variable,
  value,
  row = NULL,
  fast = F,
  write = T,
  verbose = F
)
}
\arguments{
\item{project_path}{path to project directory \code{string}}

\item{input_file}{SWAP file name to modify (required if fast=FALSE) \code{string}}

\item{output_file}{SWAP file name write (required if write=TRUE) \code{string}}

\item{variable}{SWAP variable to change \code{string}}

\item{value}{value to assign \code{string}}

\item{row}{(optional) only pass if vector (as index) or table. \code{integer}}

\item{fast}{(optional) If rswap has already parsed the swap file \code{flag}}

\item{write}{(optional) Flag to enable or disable writing of output file \code{flag}}

\item{verbose}{(optional) print status? \code{flag}}
}
\value{
Returns the path to written output file
}
\description{
This function generates a modified SWAP file, changing the \code{value} of the
passed \code{variable}. Standard behavior is to pass an \code{input_file} and an
\code{output_file}. The input file will be parsed by \code{parse_swap_file()}, the value
will be changed, and then a \code{output_file} will be written to \code{project_path}.
However, the function has a variety of different modes, see \strong{Details}
}
\details{
This function has a variety of modes. If you would like to run the function
quickly, you can enable \code{fast} which does \strong{NOT} parse the \code{input_file}.
This only works if you have already parsed the desired data, and it is
present in your \code{rswap} directory.

Another way to speed up this function is to disable \code{write}. this simply does
not write your \code{output_file}

\strong{Caution:} Not passing an \code{output_file} with \code{write=TRUE} will cause your
\code{input_file} to be overwritten!

Note: If you are changing a variable located in a SWAP vector or table, you
need to pass the \code{row} argument (which is an index for vectors, and a normal
row for tables)
}
\seealso{
\code{\link[=parse_swap_file]{parse_swap_file()}}
}
