% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{move_run_files}
\alias{move_run_files}
\title{Move run files}
\usage{
move_run_files(work_dir, project, verbose = F)
}
\arguments{
\item{work_dir}{work dir}

\item{project}{proj name}

\item{verbose}{print?}
}
\description{
Moves swap.ok and reruns.log to rswap dir
}
\keyword{internal}
