% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{NSE}
\alias{NSE}
\title{NSE}
\usage{
NSE(obs, mod)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{mod}{numeric vector of modeled values}
}
\value{
numeric NSE value rounded to 2 decimals
}
\description{
Calculates the NSE value for a paired observed and modeled vector.
}
