% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{parse_swap_file}
\alias{parse_swap_file}
\title{Parse SWAP File}
\usage{
parse_swap_file(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{swap_file}{optional swap main file name (string) defaults to "swap.swp"}

\item{verbose}{print status? (flag)}
}
\value{
Returns paths to where the parameters, vectors, and tables are written
}
\description{
This function reads in a SWAP main file writes the parameters, vectors, and
tables into csv format into the rswap directory. these can then be loaded
using \code{load_swap_parameters()}, \code{load_swap_tables()} and \code{load_swap_vectors()}
}
\details{
Also, this parsing technique was designed to work with the SWAP example cases.
If your SWAP main file structure does something different, you might run into trouble.

This function will be expanded to work with more SWAP input files like .dra
}
\seealso{
\code{\link[=clean_swap_file]{clean_swap_file()}} \code{\link[=build_rswap_directory]{build_rswap_directory()}}
}
