% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PBIAS}
\alias{PBIAS}
\title{PBIAS}
\usage{
PBIAS(obs, mod)
}
\arguments{
\item{obs}{numeric vector of observed values \strong{REQ}}

\item{mod}{numeric vector of modeled values \strong{REQ}}
}
\value{
numeric PBIAS value rounded to 2 decimal
}
\description{
Returns PBIAS value for a paired observed and modeled vector.
#TODO: change this function to accept a named vector, or dataframe.
}
