% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{plot_statistics}
\alias{plot_statistics}
\title{renamed to: "rswap_plot_performance()"}
\usage{
plot_statistics(
  project_path,
  var,
  depth = NULL,
  graph = "default",
  stat = "NSE",
  verbose = F
)
}
\arguments{
\item{project_path}{renamed to: "rswap_plot_performance()"}

\item{var}{renamed to: "rswap_plot_performance()"}

\item{depth}{renamed to: "rswap_plot_performance()"}

\item{graph}{renamed to: "rswap_plot_performance()"}

\item{stat}{renamed to: "rswap_plot_performance()"}

\item{verbose}{renamed to: "rswap_plot_performance()"}
}
\value{
renamed to: "rswap_plot_performance()"
}
\description{
renamed to: "rswap_plot_performance()"
}
