% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_swap_output}
\alias{read_swap_output}
\title{Reads SWAP output of current path}
\usage{
read_swap_output(project_path, archived = F)
}
\arguments{
\item{project_path}{String, path to the project directory / or custom save location.}

\item{archived}{Flag, needs to be true when reading from the rswap_saved archive}
}
\description{
Reads SWAP output of current path
}
