% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overunder.R
\name{ribbonize}
\alias{ribbonize}
\title{ribbionize}
\usage{
ribbonize(.data, .x, .y, .type)
}
\description{
author: Neal Grantham
source: https://www.nsgrantham.com/fill-between-two-lines-ggplot2
}
\author{
Neal Grantham
}
\keyword{internal}
