% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(obs, mod)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{mod}{numeric vector of modeled values}
}
\value{
numeric RMSE value rounded to 2 digits
}
\description{
Returns RMSE value for a paired observed and modeled vector.
}
