% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{RSR}
\alias{RSR}
\title{RSR}
\usage{
RSR(obs, mod)
}
\arguments{
\item{obs}{numeric vector of observed values}

\item{mod}{numeric vector of modeled values}
}
\value{
numeric RSR value rounded to 2 digits
}
\description{
Returns RSR value for a paired observed and modeled vector.
}
