% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{rswap_init}
\alias{rswap_init}
\title{Initialize rswap}
\usage{
rswap_init(swap_exe)
}
\arguments{
\item{swap_exe}{path to swap.exe (string)}
}
\value{
Returns the project path of "hupselbrook" SWAP setup.
}
\description{
This function is an optional starting point when using rswap for the first
time. What it does is copy in the example swap setup "hupselbrook" into a
sample project directory "rswap_init" within the same working directory as
your swap.exe executable (hence the need to pass the path to the .exe).
}
\details{
The sample directory contains the SWAP input files, as well as a template for
observed values. This provides a good starting point for setting up your own
rswap project. Just changes the name of the directory, fill in the observed
file and start modifying parameters.

If this function does not successfully complete, then there is an underlying
issue you need to fix before continuing to use rswap.
}
