% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{rswap_plot_compare}
\alias{rswap_plot_compare}
\title{Compare model runs}
\usage{
rswap_plot_compare(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to show (string)}

\item{depth}{depth of variable, if it has a depth. Otherwise leave blank (string)}

\item{verbose}{print status? (flag)}
}
\description{
This graph highlights the differences between saved model runs for a given
variable at a certain depth.
}
\seealso{
\code{\link[=rswap_plot_variable]{rswap_plot_variable()}} \code{\link[=rswap_plot_performance]{rswap_plot_performance()}} \code{\link[=rswap_plot_multi]{rswap_plot_multi()}}
}
