% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{rswap_plot_multi}
\alias{rswap_plot_multi}
\title{Plot multiple variables at once}
\usage{
rswap_plot_multi(project_path, vars, show = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{vars}{list of variables to include in the plot (string/vector)}

\item{show}{list of variables to show by default (string/vector)}

\item{verbose}{print status? (flag)}
}
\description{
Generates a 1-4 axis plot from desired variables. Designed to help
user understand how the model is working, how the variables are interacting.
}
\details{
currently supports "H", "WC", "DRAINAGE", "TEMP". Working on expanding
}
