% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{rswap_plot_performance}
\alias{rswap_plot_performance}
\title{Plot performance metrics}
\usage{
rswap_plot_performance(
  project_path,
  variable,
  stat,
  depth = NULL,
  verbose = FALSE
)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to show (string)}

\item{stat}{a statistical indicator function from package \code{hydroGOF} (string)}

\item{depth}{depth(s) to show. Leave blank for all (numeric/vector)}

\item{verbose}{print status? (flag)}
}
\description{
This function calculates the statistical performance of all the saved model
runs in the /rswap_saved/ folder, as well as the last performed run. It does
this per \code{variable} and for any/all given \code{depths}, for the passed
statistical performance indicator \code{stat} from package
\href{https://cran.r-project.org/web/packages/hydroGOF/hydroGOF.pdf}{hydroGOF}
}
\details{
The performance indicator will be the result of a function passed as a string
in the \code{stat} parameter as called from the \code{hydroGOF} package.

The following are supported for \code{stat}:

\emph{br2 cp d KGE mae md me mNSE mse nrmse NSE pbias rd rmse rNSE rPearson rSD rsr ssq VE}

Please note, passing multiple variables at differing depths will not work!
(yet)
}
\seealso{
\code{\link[=rswap_plot_variable]{rswap_plot_variable()}} \code{\link[=rswap_plot_compare]{rswap_plot_compare()}} \code{\link[=rswap_plot_multi]{rswap_plot_multi()}}
}
