% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overunder.R
\name{rswap_plot_variable}
\alias{rswap_plot_variable}
\title{Plot variable}
\usage{
rswap_plot_variable(
  project_path,
  variable,
  depth = NULL,
  interactive = F,
  verbose = F,
  filepath = NULL
)
}
\arguments{
\item{project_path}{path to project directory \code{string}}

\item{variable}{SWAP variable to plot \code{string}}

\item{depth}{depth(s) to plot \code{numeric} (optional, if left blank, all available depths will be used)}

\item{interactive}{use an interactive plotly plot? \code{flag}}

\item{verbose}{print status? \code{flag}}

\item{filepath}{if you want to save the plot to disk, provide a path. (only if interactive = FALSE)}
}
\description{
This function plots a single SWAP \code{variable} at any given \code{depth}. If no depths
are provided, all available depths are used. This plot can also be
interactive if \code{interactive = TRUE}
}
\examples{
# path to sample results (only for package internal example, you don't need
# this function)
example_path <- system.file(package = "rswap", "extdata/rswap_example_output")

# in your own projects, you would use something like this:
# example_path <- "C:/path/to/swap_folder/example_project"

rswap_plot_variable(example_path, variable = "WC")

}
\seealso{
\code{\link[=rswap_plot_compare]{rswap_plot_compare()}} \code{\link[=rswap_plot_performance]{rswap_plot_performance()}} \code{\link[=rswap_plot_multi]{rswap_plot_multi()}}
}
