% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{run_swap_scenario}
\alias{run_swap_scenario}
\title{Run SWAP Scenario}
\usage{
run_swap_scenario(scen_dir, scen_name, vars, rerun, verbose)
}
\arguments{
\item{scen_dir}{(string) path to directory containing scenario directories (must contain a \code{SWAP.exe}, see details)}

\item{scen_name}{(string) name of the directory containing the various SWAP scenarios (see details)}

\item{vars}{(vector) variables to return from output (ie. \code{vars <- c("WC_15", "drainage")})}

\item{rerun}{(logical) rerun the swap projects?}

\item{verbose}{(logical) print to console?}
}
\value{
dataframe of the scenario results for \code{var}
}
\description{
tbc
}
\details{
tbc
}
\examples{

#tbc
}
