% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_swap_format}
\alias{set_swap_format}
\title{Set SWAP Format}
\usage{
set_swap_format(value, parameter)
}
\arguments{
\item{value}{the value to be converted}

\item{parameter}{the SWAP parameter name, in string form}
}
\description{
This function converts a parameter value into the SWAP (FORTRAN) required
format. This functionality relies on SWAPtools.
}
\author{
Moritz Shore, Martin Mulder
}
