% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{set_swap_output}
\alias{set_swap_output}
\title{Set SWAP output}
\usage{
set_swap_output(parameters, variables, depths, verbose = F)
}
\arguments{
\item{parameters}{parsed parameter DF}

\item{variables}{variables desired for output}

\item{depths}{depths depth at which to output them}

\item{verbose}{print status?}
}
\value{
parameter dataframe with modified INLIST_CSV parameter.
}
\description{
by changing the inlist_csv function. if you depths for variables differ,
you need to change the "INLIST_CSV" parameter manually with change_swap_par()
}
