% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_calibration_plot.R
\name{soft_calibration_plot}
\alias{soft_calibration_plot}
\title{Soft Calibration Plot}
\usage{
soft_calibration_plot(project_path, vars, show = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{vars}{list of variables to include in the plot (string/vector)}

\item{show}{list of variables to show by default (string/vector)}

\item{verbose}{print status? (flag)}
}
\description{
Generates a 1-4 axis plot from desired variables. Designed to help
user understand how the model is working, how the variables are interacting.
}
\details{
currently supports "H", "WC", "DRAINAGE", "TEMP". Working on expanding
}
