% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{write_swap_meteo}
\alias{write_swap_meteo}
\title{Saves SWAP meteo data to project}
\usage{
write_swap_meteo(
  project_path,
  data,
  export = TRUE,
  name = NULL,
  autoset = FALSE,
  swap_file = "swap.swp",
  verbose = TRUE
)
}
\arguments{
\item{project_path}{Path to project directory (string)}

\item{data}{dataframe with SWAP meteo data as loaded by \code{load_swap_meteo()}}

\item{export}{(optional) flag, should the data be written to SWAP format?}

\item{name}{(optional) string, desired name of the meteo file.}

\item{autoset}{(optional) flag, automatically adjusts SWAP project with the
start and end times of the provided data, as well as the metfile name.}

\item{swap_file}{(optional) name of the swap file to modify (swap.swp by
default)}

\item{verbose}{(flag) print actions to console?}
}
\value{
path to where the met file was written.
}
\description{
This function writes the provided dataframe back into the rswap template
csv file, and optionally also exports the data into a SWAP compatible format
(.met). Additionally, one can automatically update the parameters \code{TSTART},
\code{TEND} and \code{METFIL} by setting \code{autoset = TRUE}.
}
\details{
Writes a / the SWAP meteo dataframe as loaded by \code{load_swap_meteo()}
}
\examples{
#tbc
}
