% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{build_rswap_directory}
\alias{build_rswap_directory}
\title{Makes a temporary sub-directory workspace for the package to run in.}
\usage{
build_rswap_directory(project_path)
}
\arguments{
\item{project_path}{string, path to project directory}
}
\description{
Makes a temporary sub-directory workspace for the package to run in.
}
\keyword{internal}
