% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{build_rswap_directory}
\alias{build_rswap_directory}
\title{Build rswap Directory}
\usage{
build_rswap_directory(project_path, force = T, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{force}{overwrite any existing directory? (flag)}

\item{verbose}{print status? (flag)}
}
\description{
Makes a temporary sub-directory workspace for the package to run SWAP in.
}
\examples{
# path to sample results (only for package internal example, you don't need
# this function)
example_path <- system.file(package = "rswap", "extdata/rswap_example_output")

# in your own projects, you would use something like this:
# example_path <- "C:/path/to/swap_folder/example_project"
# This example code will not be executed as it needs to write files!
# build_rswap_directory(example_path, verbose = TRUE)
}
