% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{build_rswap_directory}
\alias{build_rswap_directory}
\title{Build rswap Directory}
\usage{
build_rswap_directory(project_path, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{verbose}{print status? (flag)}
}
\description{
Makes a temporary sub-directory workspace for the package to run SWAP in.
}
\examples{
# path to sample results
example_path <- "./inst/extdata/rswap_example_output"

# This example code will not be executed as it needs to write files!
# build_rswap_directory(example_path, verbose = T)
}
