% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{change_swap_par}
\alias{change_swap_par}
\title{Changes a SWAP parameter}
\usage{
change_swap_par(param, name, value)
}
\arguments{
\item{param}{dataframe consisting of the parameter values}

\item{name}{name of the parameter to change}

\item{value}{that the parameter should take on}
}
\value{
modified parameter dataframe
}
\description{
passed a parameter dataframe, changes the right parameter, and returns the
dataframe
}
