% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{check_swap_message}
\alias{check_swap_message}
\title{checks SWAP error message}
\usage{
check_swap_message(status, verbose = F)
}
\arguments{
\item{msg}{file from processx}
}
\description{
checks SWAP error message
}
\keyword{internal}
