% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{clean_swp_file}
\alias{clean_swp_file}
\title{Clean swap file}
\usage{
clean_swp_file(project_path, swap_file = "swap.swp")
}
\arguments{
\item{project_path}{(REQ) (string) path where the swap file is located}

\item{swap_file}{(OPT) (string) name of the swap file (will default to "swap.swp" if left blank)}
}
\value{
cleaned up SWAP file in list form.
}
\description{
Returns a cleaned up SWAP file fit for manipulation
Make sure each table in the swap main file ends with a '* End of table' marker!!!
}
\details{
Not necessarily meant for manual use.
}
