% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{comparative_plot}
\alias{comparative_plot}
\title{this function has been renamed to: 'rswap_plot_compare()'}
\usage{
comparative_plot(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{this function has been renamed to: 'rswap_plot_compare()'}

\item{variable}{this function has been renamed to: 'rswap_plot_compare()'}

\item{depth}{this function has been renamed to: 'rswap_plot_compare()'}

\item{verbose}{this function has been renamed to: 'rswap_plot_compare()'}
}
\value{
this function has been renamed to: 'rswap_plot_compare()'
}
\description{
this function has been renamed to: 'rswap_plot_compare()'
}
