% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_plot.R
\name{comparative_plot}
\alias{comparative_plot}
\title{Comparative Plot
this graph shows the current model run compared to the previous saved
runs. It highlights the differences between changed parameters etc.}
\usage{
comparative_plot(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{(REQ) (string) path to project directory}

\item{variable}{(REQ) (string) variable to compare}

\item{depth}{(OPT) (numeric) depth of variable. leave blank if variable has
no depth}

\item{verbose}{(OPT) (logical) print status?}
}
\description{
Comparative Plot
this graph shows the current model run compared to the previous saved
runs. It highlights the differences between changed parameters etc.
}
