% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_plot.R
\name{comparative_plot}
\alias{comparative_plot}
\title{Comparative Plot}
\usage{
comparative_plot(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to show (string)}

\item{depth}{depth of variable, if it has a depth. Otherwise leave blank (string)}

\item{verbose}{print status? (flag)}
}
\description{
This graph highlights the differences between saved model runs for a given
variable at a certain depth.
}
