% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{get_depths}
\alias{get_depths}
\title{deprecated use "get_swap_depths"}
\usage{
get_depths(data, variable = NULL)
}
\arguments{
\item{data}{deprecated use "get_swap_depths"}

\item{variable}{deprecated use "get_swap_depths"}
}
\value{
deprecated use "get_swap_depths"
}
\description{
deprecated use "get_swap_depths"
}
