% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{get_performance}
\alias{get_performance}
\title{Get Model Performance}
\usage{
get_performance(
  project_path,
  archived = FALSE,
  stat = NULL,
  variable = NULL,
  depth = NULL,
  verbose = F
)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{archived}{set to \code{TRUE} if project is saved in "rswap_saved/" (flag)}

\item{stat}{statistical performance indicator (ie. "NSE") (string)}

\item{variable}{variable name. leave blank for all. (string)}

\item{depth}{depth of variable if it is depth-wise. leave blank for all (numeric)}

\item{verbose}{print status? (flag)}
}
\value{
dataframe value(s) of performance indicator(s) for given variable(s) and depth(s)
}
\description{
Returns the performance for the chosen variable(s) and performance indicator(s).
Supported performance indicators are: NSE, PBIAS, RMSE, RSR.
Supported variables are those provided in the observed data file.
}
\details{
Please note, passing multiple variables at differing depths will not work! (yet)
}
