% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_performance}
\alias{get_performance}
\title{Get model performance}
\usage{
get_performance(
  project_path,
  stat,
  variable,
  depth,
  observed_file_path = NULL,
  verbose = F
)
}
\arguments{
\item{project_path}{\strong{(REQ)} \strong{(string)}
path to the project directory / or custom save location.}

\item{stat}{(REQ) (string) either "NSE" "PBIAS" "RMSE" "RSR" or "all" for all}

\item{variable}{(REQ) (string/vector) variable name (ie. WC, H, TEMP). Use "all" for all available.}

\item{depth}{(REQ) (numeric/vector) depth of variable value (>0). Must exist in SWAP output. Use "all" for all available.}

\item{observed_file_path}{(OPT) (string) path to observed file. will default to "./project_path/observed_data.xlsx"}

\item{verbose}{(OPT) (boolean) print status to console?}
}
\value{
(dataframe) value(s) of performance indicator(s) for given variable(s) and depth(s)
}
\description{
returns the performance indicator for the chosen variable(s) and measure(s)
Supported performance indicators are: NSE, PBIAS, RMSE, RSR
Supported variables are those provided in the observed data file
Please note, passing multiple variables at differing depths will not work!
}
