% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{get_performance}
\alias{get_performance}
\title{renamed to: "get_swap_performance()"}
\usage{
get_performance(
  project_path,
  archived = FALSE,
  stat = NULL,
  variable = NULL,
  depth = NULL,
  verbose = F
)
}
\arguments{
\item{project_path}{renamed to: "get_swap_performance()"}

\item{archived}{renamed to: "get_swap_performance()"}

\item{stat}{renamed to: "get_swap_performance()"}

\item{variable}{renamed to: "get_swap_performance()"}

\item{depth}{renamed to: "get_swap_performance()"}

\item{verbose}{renamed to: "get_swap_performance()"}
}
\value{
renamed to: "get_swap_performance()"
}
\description{
renamed to: "get_swap_performance()"
}
