% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_swap_depths}
\alias{get_swap_depths}
\title{Get Depths}
\usage{
get_swap_depths(data, variable = NULL)
}
\arguments{
\item{data}{as loaded by \code{load_observed()} (dataframe)}

\item{variable}{only returns depths of the given variable when passed. leave blank for all. (string)}
}
\value{
Returns a numeric vector of depths
}
\description{
Extracts the numeric depth values found in the observed data, as loaded by
\code{load_observed()}. Can be filtered by \code{variable}.
}
\examples{
# path to sample results (only for package internal example, you don't need
# this function)
example_path <- system.file(package = "rswap", "extdata/rswap_example_output")

# in your own projects, you would use something like this:
# example_path <- "C:/path/to/swap_folder/example_project"

# load some SWAP data (either observed, or modeled using read_swap_output())
data <- load_swap_observed(project_path = example_path, verbose = TRUE)

get_swap_depths(data$data, variable = "TEMP")

}
