% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_swap_units}
\alias{get_swap_units}
\title{Get SWAP Units}
\usage{
get_swap_units(variable)
}
\arguments{
\item{variable}{SWAP variable (string)}
}
\value{
Returns unit in string from
}
\description{
returns default swap units for the given variable (Work in progress). Will
possibly also include a database on what variables are depth-wise.
}
