% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{load_observed}
\alias{load_observed}
\title{deprecated use "load_swap_observed"}
\usage{
load_observed(project_path, archived = F, verbose = F)
}
\arguments{
\item{project_path}{load_swap_observed}

\item{archived}{load_swap_observed}

\item{verbose}{load_swap_observed}
}
\value{
load_swap_observed
}
\description{
deprecated use "load_swap_observed"
}
