% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_observed}
\alias{load_observed}
\title{Load observed data (make sure to use the template .xlsx file)}
\usage{
load_observed(path, verbose = F)
}
\arguments{
\item{path}{String. Path to observed data file (.xlsx)}

\item{verbose}{Logical. Prints status reports}
}
\description{
Load observed data (make sure to use the template .xlsx file)
}
