% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{load_swap_parameters}
\alias{load_swap_parameters}
\title{Load SWAP parameters}
\usage{
load_swap_parameters(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string}

\item{swap_file}{name of the swap file to parse (string)}

\item{verbose}{print status? (flag)}
}
\value{
Tibble of parameter set
}
\description{
Loads the SWAP parameters as parsed by \code{parse_swp_file()} and returns them as
a tibble
}
