% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{load_swap_vectors}
\alias{load_swap_vectors}
\title{Load SWAP Vectors}
\usage{
load_swap_vectors(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{path to project_directory (string)}

\item{swap_file}{name of the swap file to parse (string)}

\item{verbose}{print status? (flag)}
}
\value{
Returns a list of vectors
}
\description{
This function loads all the SWAP vectors as parsed by \code{parse_swap_file()}. If
the SWAP vectors have not been parsed, then this will be done first using
\code{swap_file}
}
\seealso{
\code{\link[=parse_swap_file]{parse_swap_file()}} \code{\link[=write_swap_file]{write_swap_file()}}
}
