% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{match_mod_obs}
\alias{match_mod_obs}
\title{match mod obs}
\usage{
match_mod_obs(
  project_path,
  variable,
  observed_file_path,
  depth = NULL,
  verbose = F,
  addtional = NULL
)
}
\arguments{
\item{project_path}{req}

\item{variable}{req}

\item{observed_file_path}{req}

\item{depth}{req}

\item{verbose}{req}

\item{addtional}{req}
}
\value{
list of modelled DF and observed DF, matched
}
\description{
internal function: matches observed and modelled dataframes and returns them
}
\keyword{internal}
