% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{match_mod_obs}
\alias{match_mod_obs}
\title{deprecated, use match_swap_data()}
\usage{
match_mod_obs(project_path, variable, depth = NULL, verbose = F, archived = F)
}
\arguments{
\item{project_path}{deprecated, use match_swap_data()}

\item{variable}{deprecated, use match_swap_data()}

\item{depth}{deprecated, use match_swap_data()}

\item{verbose}{deprecated, use match_swap_data()}

\item{archived}{deprecated, use match_swap_data()}
}
\value{
deprecated, use match_swap_data()
}
\description{
deprecated, use match_swap_data()
}
