% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{melt_all_runs}
\alias{melt_all_runs}
\title{Melt all runs}
\usage{
melt_all_runs(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory}

\item{variable}{(REQ) (string) variable to be returned}

\item{depth}{(OPT) (string) depth of variable. leave blank if variable has
no depth}

\item{verbose}{print status?}
}
\value{
dataframe with columns "run" "DATE" "tag" "variable" "value"
}
\description{
Combines the past saved runs, with the current run, and the observed data
so that the combination of them is easy to plot using ggplot etc.
}
