% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{melt_all_runs}
\alias{melt_all_runs}
\title{Melt All Runs}
\usage{
melt_all_runs(project_path, variable, depth = NULL, verbose = F)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{variable}{variable to be used (string)}

\item{depth}{optional depth to filter by, leave blank for all/no depth (numeric)}

\item{verbose}{print status? (flag)}
}
\value{
Returns dataframe with columns "run" "DATE" "tag" "variable" "value"
}
\description{
This function combines the modeled and observed data of the current run,
along with all the modeled and observed data of all the saved runs
(for a single variable) and returns them in tidy format, easy to use for
plotting, etc.
}
\details{
This is once again an internal function that is made available due to it
possibly being useful for the end user.
}
