% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{parse_swp_file}
\alias{parse_swp_file}
\title{renamed to: "parse_swap_file()"}
\usage{
parse_swp_file(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{renamed to: "parse_swap_file()"}

\item{swap_file}{renamed to: "parse_swap_file()"}

\item{verbose}{renamed to: "parse_swap_file()"}
}
\value{
renamed to: "parse_swap_file()"
}
\description{
renamed to: "parse_swap_file()"
}
