% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{parse_swp_file}
\alias{parse_swp_file}
\title{Parse swap file}
\usage{
parse_swp_file(project_path, swap_file = "swap.swp", verbose = F)
}
\arguments{
\item{project_path}{(REQ) (string), path to project directory}

\item{swap_file}{(OPT) (string) name of the swap file to be used. defaults to swap.swp}

\item{verbose}{(OPT) (boolean) print status?}
}
\value{
SWAP parameters in a dataframe format and path to where the tables
were saved in a csv format (until i figure out how to return an array of
dataframes)
}
\description{
Reads the passed swap file and parses it. returns parameters and tables in a
convienient format. can also write the file directly into the rswap directory
}
\details{
Not necessarily meant for manual use.

Make sure each table in the swap main file ends with a '* End of table' marker!!!

Also this is designed to work with the same format as the example files. if
you do something different with yours,  you might run into issues
}
