% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_under.R
\name{plot_over_under}
\alias{plot_over_under}
\title{Plot over under}
\usage{
plot_over_under(
  project_path,
  observed_file_path = NULL,
  variable,
  depth = NULL,
  verbose = F
)
}
\arguments{
\item{project_path}{(REQ) (string) path to project directory}

\item{observed_file_path}{(OPT) (string) path to observed file, if it is not
in the default location}

\item{variable}{(REQ) (string) desired variable for graph}

\item{depth}{(OPT) (numeric/vector) depth at which to plot. uses all available depths if left blank.}

\item{verbose}{(OPT) (logical) print status?}
}
\description{
Plots a graph showing where the model overestimates and underestimates for
the given variable and depth(s).
}
