% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_plot.R
\name{plot_statistics}
\alias{plot_statistics}
\title{Plot statistics}
\usage{
plot_statistics(
  project_path,
  var,
  depth = NULL,
  graph = "default",
  stat = "NSE",
  custom_save_path = NULL,
  verbose = F
)
}
\arguments{
\item{project_path}{(REQ) (string) path to project directory}

\item{var}{(REQ) (string) variable to analyse}

\item{depth}{(numeric/vector) (OPT) optional, leave blank for all}

\item{graph}{(OPT) (string) either "default" "sorted" or "ggplot". leave
blank for default}

\item{stat}{(REQ) (string) Performance statistic. (NSE, PBIAS, RMSE, RSE)}

\item{custom_save_path}{(OPT) (string) pass if your saved runs have custom paths. leave blank for default}

\item{verbose}{(OPT) (boolean) print status?}
}
\description{
This function generates a plot of all the saved runs, allowing the
user to see how well their runs are performing compared to previous
attempts.
}
