% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{read_swap_output}
\alias{read_swap_output}
\title{this function has been renamed to: 'load_swap_output()'}
\usage{
read_swap_output(project_path, archived = F)
}
\arguments{
\item{project_path}{this function has been renamed to: 'load_swap_output()'}

\item{archived}{this function has been renamed to: 'load_swap_output()'}
}
\value{
this function has been renamed to: 'load_swap_output()'
}
\description{
this function has been renamed to: 'load_swap_output()'
}
