% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_swap_output}
\alias{read_swap_output}
\title{Reads SWAP output of current path}
\usage{
read_swap_output(project_path)
}
\arguments{
\item{project_path}{String, path to the project directory / or custom save location.}
}
\description{
Reads SWAP output of current path
}
