% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{rswap_plot_performance}
\alias{rswap_plot_performance}
\title{Plot performance metrics}
\usage{
rswap_plot_performance(
  project_path,
  var,
  depth = NULL,
  graph = "default",
  stat = "NSE",
  verbose = F
)
}
\arguments{
\item{project_path}{path to project directory (string)}

\item{var}{variable to show (string)}

\item{depth}{depth(s) to show. Leave blank for all (numeric/vector)}

\item{graph}{either "default" "sorted" or "ggplot" leave blank for default (string)}

\item{stat}{statistical indicator function ("NSE", "RMSE", "PBIAS", "RSR") (string)}

\item{verbose}{print status? (flag)}
}
\description{
This function calculates the statistical performance of all the saved model
runs in the /rswap_saved/ folder, as well as the last performed run. It does
this per \code{variable} and for any/all given \code{depths}, for the passed statistical
performance indicator \code{stat} (Currently supported are "NSE", "RMSE", "PBIAS", "RSR")
}
