% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overunder.R
\name{rswap_plot_variable}
\alias{rswap_plot_variable}
\title{Plot variable}
\usage{
rswap_plot_variable(
  project_path,
  variable,
  depth = NULL,
  interactive = F,
  verbose = F
)
}
\arguments{
\item{project_path}{path to project directory \code{string}}

\item{variable}{SWAP variable to plot \code{string}}

\item{depth}{depth(s) to plot \code{numeric} (optional, if left blank, all available depths will be used)}

\item{interactive}{use an interactive plotly plot? \code{flag}}

\item{verbose}{print status? \code{flag}}
}
\description{
This function plots a single SWAP \code{variable} at any given \code{depth}. If no depths
are provided, all available depths are used. This plot can also be
interactive if \code{interactive = TRUE}
}
\details{
The non-interactive version of this plot uses \code{ggbraid}, which the function
will attempt to install from github if it is not already installed.
}
\examples{
# path to sample results
example_path <- "./inst/extdata/rswap_example_output"

# Plot example path, with variable "WC" and all available depths
rswap_plot_variable(example_path, variable = "WC")
}
