% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_swap}
\alias{run_swap}
\title{Runs the SWAP model}
\usage{
run_swap(project_path)
}
\arguments{
\item{project_path}{}
}
\value{
returns name of run (change this!)
}
\description{
Runs the SWAP model
}
