% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{save_run}
\alias{save_run}
\title{deprecated: use "save_swap_run()"}
\usage{
save_run(project_path, run_name = NULL, verbose = F)
}
\arguments{
\item{project_path}{deprecated}

\item{run_name}{deprecated}

\item{verbose}{deprecated}
}
\value{
deprecated
}
\description{
deprecated: use "save_swap_run()"
}
