% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{set_swap_output}
\alias{set_swap_output}
\title{Set SWAP output}
\usage{
set_swap_output(parameters, variables, depths, verbose = F)
}
\arguments{
\item{parameters}{parameter dataframe as returned by \code{parse_swap_file()}}

\item{variables}{variables desired for output (string)}

\item{depths}{depths depth at which to output them (string)}

\item{verbose}{print status? (flag)}
}
\value{
Returns parameter dataframe with modified INLIST_CSV parameter.
}
\description{
This function alters the output parameters of your SWAP input files.
}
\details{
Currently it only modifies INLIST_CSV, but more functionality will be placed
here over time.
}
