% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{set_swap_output}
\alias{set_swap_output}
\title{Set SWAP output}
\usage{
set_swap_output(
  project_path,
  parameters,
  autoset_output = F,
  verbose = F,
  force = T
)
}
\arguments{
\item{project_path}{path to project directory}

\item{parameters}{parameter dataframe}

\item{autoset_output}{flag for matching output to observed file.}

\item{verbose}{print status? (flag)}

\item{force}{forces a reload of the SWAP project files (default, true)}
}
\value{
Returns parameter dataframe with modified INLIST_CSV parameter.
}
\description{
This function alters the output parameters of your SWAP input files.
}
\details{
Currently it only modifies INLIST_CSV, but more functionality will be placed
here over time.
}
