% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{soft_calibration_plot}
\alias{soft_calibration_plot}
\title{this function has been renamed to: 'rswap_plot_multi()'}
\usage{
soft_calibration_plot(project_path, vars, show = NULL, verbose = F)
}
\arguments{
\item{project_path}{this function has been renamed to: 'rswap_plot_multi()'}

\item{vars}{this function has been renamed to: 'rswap_plot_multi()'}

\item{show}{this function has been renamed to: 'rswap_plot_multi()'}

\item{verbose}{this function has been renamed to: 'rswap_plot_multi()'}
}
\value{
this function has been renamed to: 'rswap_plot_multi()'
}
\description{
this function has been renamed to: 'rswap_plot_multi()'
}
