% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_calibration_plot.R
\name{soft_calibration_plot}
\alias{soft_calibration_plot}
\title{Soft Calibration Plot}
\usage{
soft_calibration_plot(project_path, vars, show = NULL)
}
\arguments{
\item{project_path}{path to project directory}

\item{vars}{(REQ) (string/vector) list of variables to include in the plot}

\item{show}{(OPT) (string/vector) list of variables to show by default}
}
\description{
Generates a 1-4 axis plot from desired variables. Designed to help
user understand how the model is working.
}
\details{
currently supports "H", "WC", "DRAINAGE", "TEMP"
}
