% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_support.R
\name{update_swp_paths}
\alias{update_swp_paths}
\title{deprecated, use 'update_swap_paths()' instead}
\usage{
update_swp_paths(project_path, swap_exe, parameters, verbose = F)
}
\arguments{
\item{project_path}{deprecated}

\item{swap_exe}{deprecated}

\item{parameters}{deprecated}

\item{verbose}{deprecated}
}
\value{
deprecated
}
\description{
deprecated, use 'update_swap_paths()' instead
}
