% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{update_swp_paths}
\alias{update_swp_paths}
\title{updates the file paths in the swap main file}
\usage{
update_swp_paths(temp_directory, swap_file, swap_exe, verbose)
}
\arguments{
\item{temp_directory}{path to the temp directory}

\item{swap_file}{name of the swap file to be modified}

\item{swap_exe}{path to swap.exe}

\item{verbose}{print status?}
}
\description{
updates the file paths in the swap main file
}
\keyword{internal}
