% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{update_swp_paths}
\alias{update_swp_paths}
\title{Update swp main file paths}
\usage{
update_swp_paths(project_path, swap_exe, parameters, verbose)
}
\arguments{
\item{project_path}{path to the temp directory}

\item{swap_exe}{path to swap.exe}

\item{parameters}{SWAP main file parameters in dataframe format}

\item{verbose}{print status?}
}
\value{
swap parameter dataframe
}
\description{
updates the file paths in the swap main file
}
\keyword{internal}
