% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{update_swp_paths}
\alias{update_swp_paths}
\title{Update SWAP Main File Paths}
\usage{
update_swp_paths(project_path, swap_exe, parameters, verbose = F)
}
\arguments{
\item{project_path}{path to the temp directory (string)}

\item{swap_exe}{path to swap.exe (string)}

\item{parameters}{SWAP main file parameters (string)}

\item{verbose}{print status? (flag)}
}
\value{
Returns the SWAP parameter dataframe with modified path values
}
\description{
As \code{rswap} creates a new directory for the model file, the paths in the those
files need to be update to still be correct. This function does just that.
}
