% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{write_swap_file}
\alias{write_swap_file}
\title{write swap file}
\usage{
write_swap_file(
  project_path,
  parameters,
  table_path,
  outpath = NULL,
  outfile = "rswap.swp",
  verbose = F
)
}
\arguments{
\item{project_path}{(REQ) (string) path to project directory}

\item{parameters}{(REQ) (dataframe) dataframe of project parameters as created
by the parse_swap_file() function}

\item{table_path}{(REQ) (string) path to swap tables stored as csv as written
by the parse_swap_fil function}

\item{outpath}{(OPT) (string) custom write location, defaults to project_directory/rswap/}

\item{outfile}{(OPT) (string) custom file name. defaults to rswap.swp}

\item{verbose}{(OPT) (boolean) print status?}
}
\description{
writes the swap file from the passed parameter dataframe, and the passed path
to the tables stored as csv from the parse_swap_file() function
}
