% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_parameters.R
\name{write_swap_parameters}
\alias{write_swap_parameters}
\title{Write SWAP parameters}
\usage{
write_swap_parameters(project_path, parameters, type = "main", verbose = F)
}
\arguments{
\item{project_path}{path to project directory}

\item{parameters}{parameter set as loaded by \code{load_swap_parameters()} (dataframe)}

\item{type}{supported swap parameter sets, current includes "main" (default) and "dra" (drainage)}

\item{verbose}{print status? (flag)}
}
\description{
Writes the SWAP parameter set into the rswap directory in .csv format
}
\seealso{
\code{\link[=load_swap_parameters]{load_swap_parameters()}} \code{\link[=write_swap_tables]{write_swap_tables()}} \code{\link[=write_swap_vectors]{write_swap_vectors()}}
}
